Move TranslationForWorkSheet file to worksheets map
