+++ frontend/src/worksheets/TableListOfWorkSheets.js
@@ -7,7 +7,7 @@ import Button from '../Button'
 import {
   typeOfWorkTranslation,
   statusTranslation,
-} from '../TranslationForWorkSheet'
+} from './TranslationForWorkSheet'
 import useCurrentUser from '../hooks/useCurrentUser'
 import CloseButton from '../specialButtons/CloseButton'
 import { closeWorkSheet, finalizeWorkSheet } from '../api/WorkSheetAPI'
